Implement WASI.getImportObject and WASI.initialize in node:wasi (#12755, #28534)
